Make empty values in the config table not set properties.
If the input for "key" is blank, then the entire row is skipped now,
and if an input for a non-"key" column is blank, that paricular property
is left unset on the config object. This allows Cohoard to properly use
a fallback value, instead of thinking that the property is set to the
empty string.

Fixes #12