Make the name field use a fallback value if not provided.
Previously, it was possible to define a User in a config to not have a
name. This is somewhat annoying, since you almost always want a User to
have a name in any template. We fix this by having the name fall back to
the key, which is always provided.

This also lets us simplify templates. Now `{{ post.user.name | default (value = post.user.key) }}`
can become just `{{ post.user.name }}`

Fixes #19.