Just kidding, don't use wee-alloc
`wee-alloc` causes a strange error somewhere within the generated
wasm. It looks like the crate is unmaintained, and also the size savings
are like 2 kB--not enough the hassle of debugging it.

(The error I is below)
```
Uncaught TypeError: Error resolving module specifier “env”. Relative module specifiers must start with “./”, “../” or “/”.
```