Remove `load_config` and merge error message divs together.
`load_config` is useless, since we can just access the current config
from the ConfigTable class. Instead of loading the config explicitly, we
do it at render time. This might be marginally slower but probably not
that much slower.

Also, since we don't have a `load_config` function, the error reporting
associated with that function has been merged with the `render` function
error reporting.